Get sites if sites_last_request is None, set last_update_time after getting sites
It was possible to set the last update time for sites, without getting the sites
if the API call failed. If the sites were requested again, then None is returned
as the code thinks the sites have been updated. This commit avoids this by: a)
always getting the sites if sites_last_request is None; b) Update
last_update_time after the sites have been retrieved from the API